requirements.txt
openai>=1.0.0
flask>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pdf2image>=1.16.0
Pillow>=10.0.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
pandas>=2.0.0
tiktoken>=0.5.0

